requirements.txt
gradio==5.44.1

langchain-core==0.3.75
langchain-google-genai==2.1.10

python-dotenv==1.1.1
